Cleanup NoEpochsConsensus tests
